+++ src/calendar/day/index.tsx
@@ -9,7 +9,7 @@ import React, {Component} from 'react';
 // @ts-expect-error
 import {shouldUpdate} from '../../component-updater';
 // @ts-expect-error
-import {isToday as dateutils_isToday} from '../../dateutils';
+import {formatNumbers, isToday as dateutils_isToday} from '../../dateutils';
 // @ts-expect-error
 import {xdateToData} from '../../interface';
 // @ts-expect-error
@@ -17,7 +17,6 @@ import {SELECT_DATE_SLOT} from '../../testIDs';
 import BasicDay, {BasicDayProps} from './basic';
 import PeriodDay from './period';
 import {MarkingProps} from './marking';
-
 
 const basicDayPropsTypes = omit(BasicDay.propTypes, 'date');
 
@@ -117,7 +116,7 @@ export default class Day extends Component<DayProps> {
         testID={`${SELECT_DATE_SLOT}-${date.dateString}`}
         accessibilityLabel={accessibilityLabel}
       >
-        {date ? day?.getDate() : day}
+        {formatNumbers(date ? day?.getDate() : day)}
       </Component>
     );
   }