+++ src/day-state-manager.js
@@ -1,7 +1,7 @@
 const {isToday, isDateNotInTheRange, sameMonth} = require('./dateutils');
 const {parseDate, toMarkingFormat} = require('./interface');
 
-function getState(day, current, props) {
+export function getState(day, current, props) {
   const {minDate, maxDate, disabledByDefault, context} = props;
   const _minDate = parseDate(minDate);
   const _maxDate = parseDate(maxDate);
@@ -22,5 +22,3 @@ function getState(day, current, props) {
 
   return state;
 }
-
-export {getState};